fix (test): faulty test case
in previous versions, it was possible to refresh your access token within the refresh test case and get a different access token.
this was faulty behaviour as it relied on creating access tokens with expiry with time resolution in miliseconds, as opposed to seconds (which are needed for the JWT)
removing this might result in a regression further down the line which is not caught by the test case (refresh call not actually refreshing the access token), but the only way to catch it automatically would be to slow down the test by a second (at least), which I am not willing to do.