feat(routes): generic endpoints
Introduces an endpoint generic, allowing for standardized request
parsing, relieving router code from a bunch of filler code. This opens
the door to OpenAPI document generator implementations in the near
future.

The commit is not fully compatible with previous commits and the API has
been changed in a few instances (ex. ?start -> ?range_start).